Traits Database Cleanup
Start removing bad references in Traits Database.
If a trait is provided by a Top Level Race (Elf, Halfing, etc.) we
don't want to call out a Subrace. Similarly, if a trait is provided
by a Subrace, we don't want to call out a Race.

TODO: Maybe just merge Race and Subrace lists?